fix(type): Cut the circular module dependency for some types
