+++ src/telegram/actions/donate.ts
@@ -15,11 +15,12 @@ import { Logger } from "../../logger/index.js";
 import { collectAnalytics } from "../../analytics/index.js";
 import { donationLevels } from "../../const.js";
 import { toCurrency } from "../../text/utils.js";
-import type { TgInlineKeyboardButton, TgMessage } from "../api/types.js";
+import type { TgMessage } from "../api/types.js";
 import type { PaymentService } from "../../donate/types.js";
 import type { AnalyticsData } from "../../analytics/ga/types.js";
 import type { LanguageCode } from "../../recognition/types.js";
 import type { ChatId, MessageThreadId } from "../api/core.js";
+import type { TgInlineKeyboardButton } from "../api/groups/chats/chats-types.js";
 
 const getDonateButton = (
   price: number,
