+++ src/telegram/types.ts
@@ -15,7 +15,7 @@ import {
 } from "./helpers.js";
 import { Logger } from "../logger/index.js";
 import { getTranslator } from "../text/index.js";
-import { type TgMessage, type TgMessageOptions } from "./api/types.js";
+import { type TgMessage } from "./api/types.js";
 import { type AnalyticsData } from "../analytics/ga/types.js";
 import type { LanguageCode } from "../recognition/types.js";
 import type { ValueOf } from "../common/types.js";
@@ -25,6 +25,7 @@ import {
   type MessageThreadId,
   TgChatId,
 } from "./api/core.js";
+import type { TgMessageOptions } from "./api/groups/chats/chats-types.js";
 
 export const VoiceContentReason = {
   Ok: "Ok",
